requirements.txt
requests
pandas
matplotlib
mplfinance
python-dotenv